+++ examples/resources/meraki_appliance_vlans/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_appliance_vlans.example "<organization_id>,<network_id>"